chore: move constants to dapp-config
