+++ tools/scripts/utils/on-chain-utils/fetchNative.ts
@@ -1,15 +1,13 @@
 import { ApiPromise } from '@polkadot/api';
-import { ICurrency } from '@webb-tools/dapp-config/src/on-chain-config/on-chain-config-base';
-import { zeroAddress } from '@webb-tools/dapp-types';
-import { parseTypedChainId } from '@webb-tools/sdk-core';
-import { ethers } from 'ethers';
-
-import chainData from './fixtures/native.json';
 import {
-  DEFAULT_EVM_CURRENCY,
   LOCALNET_CHAIN_IDS,
   SELF_HOSTED_CHAIN_IDS,
-} from './shared';
+} from '@webb-tools/dapp-config/src/chains';
+import { DEFAULT_EVM_CURRENCY } from '@webb-tools/dapp-config/src/currencies';
+import { ICurrency } from '@webb-tools/dapp-config/src/types';
+import { parseTypedChainId } from '@webb-tools/sdk-core';
+import { ZERO_ADDRESS } from '@webb-tools/utils';
+import chainData from './fixtures/native.json';
 import { DEFAULT_NATIVE_INDEX } from './shared';
 
 /// Private Methods
@@ -34,7 +32,7 @@ async function fetchEVMNativeCurrency(
 
   return {
     ...chain.nativeCurrency,
-    address: zeroAddress,
+    address: ZERO_ADDRESS,
   };
 }
