+++ tools/scripts/utils/on-chain-utils/getViemClient.ts
@@ -1,18 +1,15 @@
+import { LOCALNET_CHAIN_IDS } from '@webb-tools/dapp-config/src/chains';
 import { chainsConfig } from '@webb-tools/dapp-config/src/chains/chain-config';
 import { EVMChainId } from '@webb-tools/dapp-types/src/EVMChainId';
 import { parseTypedChainId } from '@webb-tools/sdk-core';
-import { Chain } from 'viem';
-import { createPublicClient, http } from 'viem';
+import { Chain, createPublicClient, http } from 'viem';
 
 import * as chains from 'viem/chains';
 
 // At the time of writing, Viem does not support multicall for these chains.
-const VIEM_NOT_SUPPORTED_MULTICALL_CHAINS = [
-  EVMChainId.ScrollAlpha,
-  EVMChainId.AthenaLocalnet,
-  EVMChainId.HermesLocalnet,
-  EVMChainId.DemeterLocalnet,
-];
+const VIEM_NOT_SUPPORTED_MULTICALL_CHAINS = [EVMChainId.ScrollAlpha].concat(
+  LOCALNET_CHAIN_IDS
+);
 
 /**
  * Gets the chain object for the given chain id.
